api/requirements.txt
djangorestframework==3.1.1
django-oauth-toolkit==0.8.0
djangorestframework-jwt==1.4.0
